requirements.txt
Flask
slack-sdk
openai
faiss-cpu
pdfplumber
requests
python-dotenv
numpy

